@@ -6,15 +6,14 @@ version: 1.7.5+480
 
 environment:
   sdk: ">=2.17.1 <3.0.0"
-  # sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^1.1.0
-  protobuf: ^3.0.0
-  uuid: ^3.0.4
-  package_info_plus: ^4.0.2
+  http: ^1.4.0
+  protobuf: ^4.1.0
+  uuid: ^4.5.1
+  package_info_plus: ^8.3.0
   recase: ^4.1.0
   # share: ^0.6.5+4
   flutter_reorderable_list: ^1.2.0
@@ -22,34 +21,32 @@ dependencies:
   animated_list_plus: ^0.5.0
   flutter_keyboard_visibility: ^6.0.0
   # aeyrium_sensor: ^1.0.7
-  sensors: ^2.0.3
+  # sensors: ^2.0.3
   # dart_midi: 1.0.1
   quiver: ^3.0.1
   path_drawing: 1.0.1
   path_provider: ^2.0.2
   shared_preferences: ^2.0.6
   base_x: ^2.0.0
-  archive: ^3.1.2
+  archive: ^4.0.7
   font_awesome_flutter: ^10.5.0
   cupertino_icons: ^1.0.5
-  fluro: "^2.0.3"
+  fluro: ^2.0.3
   # flutter_appavailability: ^0.0.21
-  native_device_orientation: ^1.1.4
-  webview_flutter: ^4.2.2
+  native_device_orientation: ^2.0.3
+  webview_flutter: ^4.13.0
   # flutter_icons: ^1.1.0
   # flutter_blue: ^0.8.0
-  #  path: ../flutter_blue
   # flutter_reactive_ble: #^3.1.0
-  #   path: ../flutter_reactive_ble
-  flutter_midi_command: ^0.4.11
-  #  path: ../FlutterMidiCommand
+  flutter_midi_command: ^0.5.3
   dcache: ^0.4.0
   provider: ^6.0.2
-  matrix4_transform: ^2.0.1
+  matrix4_transform: ^4.0.1
   share_plus: ^11.0.0
   collection: ^1.19.1
   appcheck: ^1.5.4+1
   material_design_icons_flutter: ^7.0.7296
+  sensors_plus: ^6.1.1
 
 dev_dependencies:
   flutter_test:
