@@ -30,7 +30,7 @@ class MusicPreviewRenderer {
       required this.renderSections,
       required this.renderPartNames,
       required this.musicViewMode,
-      this.renderColor});
+      required this.renderColor});
 
   MusicSystemPainter get painter {
     final parts = score.parts;
@@ -51,6 +51,9 @@ class MusicPreviewRenderer {
       partTopOffsets: ValueNotifier(partTopOffsets),
       staffOffsets: ValueNotifier(staffOffsets),
       colorGuideOpacityNotifier: ValueNotifier(0),
+      tappedBeat: ValueNotifier(null),
+      tappedPart: ValueNotifier(null),
+      bluetoothControllerPressedNotes: ValueNotifier(Map()),
       colorblockOpacityNotifier: ValueNotifier(
           AppSettings.globalRenderingMode == RenderingMode.colorblock ? 1 : 0),
       notationOpacityNotifier: ValueNotifier(
@@ -76,7 +79,7 @@ class MusicPreviewRenderer {
   double get maxWidth =>
       (extraBeatsSpaceForClefs + score.beatCount) * beatWidth * scale;
   double get actualWidth => min(maxWidth, width);
-  Future<ui.Image> get renderedScoreImage async {
+  Future<ui.Image?> get renderedScoreImage async {
     if (height < 1 || width < 1) {
       return null;
     }
@@ -105,9 +108,9 @@ class MusicPreviewRenderer {
     return data;
   }
 
-  Future<Uint8List> get renderedScoreImageData async {
+  Future<Uint8List?> get renderedScoreImageData async {
     final image = await renderedScoreImage;
     return Uint8List.sublistView(
-        await image.toByteData(format: ui.ImageByteFormat.png));
+        (await image?.toByteData(format: ui.ImageByteFormat.png))!);
   }
 }
